+++ src/main/java/shadows/apotheosis/adventure/AdventureModule.java
@@ -71,8 +71,8 @@ public class AdventureModule {
 
 	@SubscribeEvent
 	public void preInit(ApotheosisConstruction e) {
-		ObfuscationReflectionHelper.setPrivateValue(RangedAttribute.class, (RangedAttribute) Attributes.ARMOR, 40D, "maxValue");
-		ObfuscationReflectionHelper.setPrivateValue(RangedAttribute.class, (RangedAttribute) Attributes.ARMOR_TOUGHNESS, 30D, "maxValue");
+		ObfuscationReflectionHelper.setPrivateValue(RangedAttribute.class, (RangedAttribute) Attributes.ARMOR, 40D, "f_22308_");
+		ObfuscationReflectionHelper.setPrivateValue(RangedAttribute.class, (RangedAttribute) Attributes.ARMOR_TOUGHNESS, 30D, "f_22308_");
 	}
 
 	@SubscribeEvent