+++ src/main/java/shadows/apotheosis/adventure/loot/LootRarity.java
@@ -143,7 +143,10 @@ public record LootRarity(String id, TextColor color, List<LootRule> rules, int o
 					return;
 				}
 				List<Affix> available = AffixHelper.byType(type).stream().filter(a -> a.canApplyTo(stack, rarity) && !currentAffixes.contains(a)).collect(Collectors.toList());
-				if (available.size() == 0) AdventureModule.LOGGER.error("Failed to execute LootRule {}/{}!", type, chance);
+				if (available.size() == 0) {
+					AdventureModule.LOGGER.error("Failed to execute LootRule {}/{}/{}/{}!", LootCategory.forItem(stack), rarity.id(), type, chance);
+					return;
+				}
 				Collections.shuffle(available, rand);
 				currentAffixes.add(available.get(0));
 			}